+++ src/app/componentRegistry.js
@@ -8,7 +8,7 @@ import HoverableSpatialSubscriber from '../components/spatial/HoverableSpatialSu
 import HoverableHeatmapSubscriber from '../components/heatmap/HoverableHeatmapSubscriber';
 import LayerControllerSubscriber from '../components/layer-controller/LayerControllerSubscriber';
 import HiGlassSubscriber from '../components/higlass/HiGlassSubscriber';
-import CellSetSizePlotSubscriber from '../components/sets/CellSetSizePlotSubscriber';
+import CellSetSizesPlotSubscriber from '../components/sets/CellSetSizesPlotSubscriber';
 
 const registry = {
   description: DescriptionSubscriber,
@@ -21,7 +21,7 @@ const registry = {
   heatmap: HoverableHeatmapSubscriber,
   layerController: LayerControllerSubscriber,
   higlass: HiGlassSubscriber,
-  cellSetSize: CellSetSizePlotSubscriber,
+  cellSetSizes: CellSetSizesPlotSubscriber,
 };
 
 export default function getComponent(name) {
